+++ CHANGES.md
@@ -12,6 +12,12 @@
 - Semicolon in parameter value can now be escaped with `\:` [#190](https://github.com/MarketSquare/robotframework-tidy/issues/190)
 - Default separator can be changed from space to tabular with new ``--separator`` option [#184](https://github.com/MarketSquare/robotframework-tidy/issues/184)
 
+### Fixes
+- Robotidy will not print "Loaded configuration from ... " if the configuration is empty [#193](https://github.com/MarketSquare/robotframework-tidy/issues/193)
+
+### Other
+- ReplaceRunKeywordIf now removes ELSE branch if it is unnecessary [#192](https://github.com/MarketSquare/robotframework-tidy/issues/192)
+
 ## 1.5.1
 
 ### Fixes